+++ common/mcu_utils.h
@@ -25,7 +25,7 @@
 #include <linux/types.h>
 #include <linux/io.h>
 
-int memcpy_toio_32(void *pdst, const void *psrc, size_t size);
-int memcpy_fromio_32(void *pdst, void *psrc, size_t size);
+int memcpy_toio_32(void *pdst, const void *psrc, int size);
+int memcpy_fromio_32(void *pdst, void *psrc, int size);
 
 #endif /* _MCU_UTILS_H_ */
